+++ modules/AWSPowerShell/Cmdlets/AmazonCloudWatch/Basic/Write-CWMetricData-Cmdlet.cs
@@ -35,17 +35,29 @@ namespace Amazon.PowerShell.Cmdlets.CW
     /// 
     ///  
     /// <para>
-    /// Each <code>PutMetricData</code> request is limited to 8 KB in size for HTTP GET requests
-    /// and is limited to 40 KB in size for HTTP POST requests.
+    /// Each <code>PutMetricData</code> request is limited to 40 KB in size for HTTP POST
+    /// requests.
     /// </para><para>
     /// Although the <code>Value</code> parameter accepts numbers of type <code>Double</code>,
     /// Amazon CloudWatch rejects values that are either too small or too large. Values must
     /// be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base
     /// 2). In addition, special values (e.g., NaN, +Infinity, -Infinity) are not supported.
     /// </para><para>
+    /// You can use up to 10 dimensions per metric to further clarify what data the metric
+    /// collects. For more information on specifying dimensions, see <a href="http://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/publishingMetrics.html">Publishing
+    /// Metrics</a> in the <i>Amazon CloudWatch User Guide</i>.
+    /// </para><para>
     /// Data points with time stamps from 24 hours ago or longer can take at least 48 hours
     /// to become available for <a>GetMetricStatistics</a> from the time they are submitted.
-    /// </para>
+    /// </para><para>
+    /// CloudWatch needs raw data points to calculate percentile statistics. If you publish
+    /// data using a statistic set instead, you cannot retrieve percentile statistics for
+    /// this data unless one of the following conditions is true:
+    /// </para><ul><li><para>
+    /// The SampleCount of the statistic set is 1
+    /// </para></li><li><para>
+    /// The Min and the Max of the statistic set are equal
+    /// </para></li></ul>
     /// </summary>
     [Cmdlet("Write", "CWMetricData", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("None","System.String")]