+++ modules/AWSPowerShell/Cmdlets/AmazonElasticLoadBalancingV2/Basic/Get-ELB2SSLPolicy-Cmdlet.cs
@@ -32,7 +32,8 @@ namespace Amazon.PowerShell.Cmdlets.ELB2
     /// 
     ///  
     /// <para>
-    /// Note that the only supported policy at this time is ELBSecurityPolicy-2015-05.
+    /// For more information, see <a href="http://docs.aws.amazon.com/elasticloadbalancing/latest/application/create-https-listener.html#describe-ssl-policies">Security
+    /// Policies</a> in the <i>Application Load Balancers Guide</i>.
     /// </para><br/><br/>This operation automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output.
     /// </summary>
     [Cmdlet("Get", "ELB2SSLPolicy")]