+++ modules/AWSPowerShell/Cmdlets/AmazonElasticLoadBalancingV2/Basic/New-ELB2TargetGroup-Cmdlet.cs
@@ -129,8 +129,9 @@ namespace Amazon.PowerShell.Cmdlets.ELB2
         #region Parameter Matcher_HttpCode
         /// <summary>
         /// <para>
-        /// <para>The HTTP codes. The default value is 200. You can specify multiple values (for example,
-        /// "200,202") or a range of values (for example, "200-299").</para>
+        /// <para>The HTTP codes. You can specify values between 200 and 499. The default value is 200.
+        /// You can specify multiple values (for example, "200,202") or a range of values (for
+        /// example, "200-299").</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]
@@ -140,7 +141,9 @@ namespace Amazon.PowerShell.Cmdlets.ELB2
         #region Parameter Name
         /// <summary>
         /// <para>
-        /// <para>The name of the target group.</para>
+        /// <para>The name of the target group.</para><para>This name must be unique per region per account, can have a maximum of 32 characters,
+        /// must contain only alphanumeric characters or hyphens, and must not begin or end with
+        /// a hyphen.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]